Extract internal/device package from internal/sensors (#105)
* refactor: extract internal/device package from internal/sensors

Moves Device model, DeviceStore interface, postgres implementation, and
DeviceService into internal/device. Sensors package keeps thin wrappers
for backward compat. Platform middleware now uses device.Info/ContextKey.

Closes #97

* Extract internal/peripheral and internal/measurement packages (#106)

* refactor: extract internal/peripheral and internal/measurement packages

Moves peripheral types, store, and service into internal/peripheral.
Introduces internal/measurement with Point/Query/Reading/Writer/Querier
so peripheral and sensors can share measurement types without circular
imports. InfluxDB client updated to implement measurement.Client.
Sensors package keeps thin aliases/wrappers for backward compat.

Closes #98

* Move InfluxDB store and ReadingsService into internal/measurement (#107)

* refactor: move InfluxDB store and ReadingsService into internal/measurement

Moves the InfluxDB Client implementation into internal/measurement/store_influx.go
and ReadingsService into internal/measurement/service.go. A DeviceFinder interface
decouples measurement from device, satisfied by a thin bridge in sensors.
sensors/influx.go and sensors/readings_service.go become thin wrappers.

Closes #99

* Extract internal/provisioning package from internal/sensors (#108)

* refactor: extract internal/provisioning package from internal/sensors

Moves provisioning store, postgres implementation, errors, and service
into internal/provisioning. Sensors package keeps thin aliases/wrappers
for backward compat.

Closes #100

* refactor: split provisioning.go into errors, store, service files


---------


* refactor: move ErrInfluxWrite to errors.go in measurement package


---------


* refactor: move peripheral error sentinels to errors.go


---------


* refactor: move device error sentinels to errors.go


---------
