+++ internal/measurement/errors.go
@@ -1,0 +1,5 @@
+package measurement
+
+import "errors"
+
+var ErrInfluxWrite = errors.New("failed to persist reading")
