+++ internal/peripheral/errors.go
@@ -1,0 +1,11 @@
+package peripheral
+
+import "errors"
+
+var (
+	ErrNotAnActuator  = errors.New("peripheral is not an actuator")
+	ErrNotFound       = errors.New("peripheral not found")
+	ErrAlreadyExists  = errors.New("peripheral already exists")
+	ErrPinInUse       = errors.New("peripheral pin already in use")
+	ErrInvalidCommand = errors.New("action must be 'set' or 'schedule'")
+)
