+++ internal/sensors/provisioning_service.go
@@ -1,29 +1,14 @@
 package sensors
 
 import (
-	"context"
 	"log/slog"
+
+	"github.com/fishhub-oss/fishhub-server/internal/provisioning"
 )
 
-// ProvisioningService orchestrates device provisioning from the user side.
-type ProvisioningService struct {
-	store  ProvisioningStore
-	logger *slog.Logger
-}
+// ProvisioningService is an alias for provisioning.Service kept for backward compatibility.
+type ProvisioningService = provisioning.Service
 
 func NewProvisioningService(store ProvisioningStore, logger *slog.Logger) *ProvisioningService {
-	if logger == nil {
-		logger = slog.Default()
-	}
-	return &ProvisioningService{store: store, logger: logger}
-}
-
-// Provision returns an existing unused provisioning code for the user or
-// creates a new one.
-func (s *ProvisioningService) Provision(ctx context.Context, userID string) (code string, err error) {
-	code, err = s.store.GetOrCreateCode(ctx, userID)
-	if err != nil {
-		s.logger.Error("provision: get or create code", "user_id", userID, "error", err)
-	}
-	return code, err
+	return provisioning.NewService(store, logger)
 }
