+++ internal/sensors/store_postgres.go
@@ -1,137 +1,13 @@
 package sensors
 
 import (
-	"context"
 	"database/sql"
-	"errors"
-	"fmt"
+
+	"github.com/fishhub-oss/fishhub-server/internal/device"
 )
 
-type postgresDeviceStore struct {
-	db *sql.DB
-}
-
+// NewDeviceStore returns a device.Store backed by Postgres.
+// Kept here for backward compatibility with callers that import sensors.
 func NewDeviceStore(db *sql.DB) DeviceStore {
-	return &postgresDeviceStore{db: db}
+	return device.NewStore(db)
 }
-
-func (s *postgresDeviceStore) ListByUserID(ctx context.Context, userID string) ([]Device, error) {
-	rows, err := s.db.QueryContext(ctx, `
-		SELECT id, COALESCE(name, ''), created_at
-		FROM devices
-		WHERE user_id = $1 AND deleted_at IS NULL
-		ORDER BY created_at DESC
-	`, userID)
-	if err != nil {
-		return nil, fmt.Errorf("list devices: %w", err)
-	}
-	defer rows.Close()
-
-	devices := []Device{}
-	for rows.Next() {
-		var d Device
-		if err := rows.Scan(&d.ID, &d.Name, &d.CreatedAt); err != nil {
-			return nil, fmt.Errorf("scan device: %w", err)
-		}
-		devices = append(devices, d)
-	}
-	return devices, rows.Err()
-}
-
-func (s *postgresDeviceStore) FindByID(ctx context.Context, deviceID string) (Device, error) {
-	var d Device
-	err := s.db.QueryRowContext(ctx, `
-		SELECT id, user_id, COALESCE(name, ''), created_at
-		FROM devices
-		WHERE id = $1 AND deleted_at IS NULL
-	`, deviceID).Scan(&d.ID, &d.UserID, &d.Name, &d.CreatedAt)
-	if errors.Is(err, sql.ErrNoRows) {
-		return Device{}, ErrDeviceNotFound
-	}
-	if err != nil {
-		return Device{}, fmt.Errorf("find device by id: %w", err)
-	}
-	return d, nil
-}
-
-func (s *postgresDeviceStore) FindByIDAndUserID(ctx context.Context, deviceID, userID string) (Device, error) {
-	var d Device
-	err := s.db.QueryRowContext(ctx, `
-		SELECT id, user_id, COALESCE(name, ''), created_at
-		FROM devices
-		WHERE id = $1 AND user_id = $2 AND deleted_at IS NULL
-	`, deviceID, userID).Scan(&d.ID, &d.UserID, &d.Name, &d.CreatedAt)
-	if errors.Is(err, sql.ErrNoRows) {
-		return Device{}, ErrDeviceNotFound
-	}
-	if err != nil {
-		return Device{}, fmt.Errorf("find device: %w", err)
-	}
-	return d, nil
-}
-
-func (s *postgresDeviceStore) DeleteDevice(ctx context.Context, deviceID, userID string) (string, error) {
-	var mqttUsername string
-	err := s.db.QueryRowContext(ctx, `
-		UPDATE devices
-		SET deleted_at = now()
-		WHERE id = $1 AND user_id = $2 AND deleted_at IS NULL
-		RETURNING COALESCE(mqtt_username, '')
-	`, deviceID, userID).Scan(&mqttUsername)
-	if errors.Is(err, sql.ErrNoRows) {
-		return "", ErrDeviceNotFound
-	}
-	if err != nil {
-		return "", fmt.Errorf("delete device: %w", err)
-	}
-	return mqttUsername, nil
-}
-
-func (s *postgresDeviceStore) GetActivationStatus(ctx context.Context, deviceID string) (ActivationStatus, error) {
-	var username, password sql.NullString
-	var pendingOutbox bool
-	err := s.db.QueryRowContext(ctx, `
-		SELECT
-			d.mqtt_username,
-			d.mqtt_password,
-			EXISTS (
-				SELECT 1 FROM outbox_events
-				WHERE payload->>'device_id' = $1::text
-				  AND status IN ('pending', 'processing')
-			)
-		FROM devices d
-		WHERE d.id = $1::uuid AND d.deleted_at IS NULL
-	`, deviceID).Scan(&username, &password, &pendingOutbox)
-	if errors.Is(err, sql.ErrNoRows) {
-		return ActivationStatus{}, ErrDeviceNotFound
-	}
-	if err != nil {
-		return ActivationStatus{}, fmt.Errorf("get activation status: %w", err)
-	}
-	if !username.Valid || !password.Valid || pendingOutbox {
-		return ActivationStatus{Ready: false}, nil
-	}
-	return ActivationStatus{
-		Ready:        true,
-		MQTTUsername: username.String,
-		MQTTPassword: password.String,
-	}, nil
-}
-
-func (s *postgresDeviceStore) PatchDevice(ctx context.Context, deviceID, userID, name string) (Device, error) {
-	var d Device
-	err := s.db.QueryRowContext(ctx, `
-		UPDATE devices
-		SET name = $1
-		WHERE id = $2 AND user_id = $3 AND deleted_at IS NULL
-		RETURNING id, user_id, COALESCE(name, ''), created_at
-	`, name, deviceID, userID).Scan(&d.ID, &d.UserID, &d.Name, &d.CreatedAt)
-	if errors.Is(err, sql.ErrNoRows) {
-		return Device{}, ErrDeviceNotFound
-	}
-	if err != nil {
-		return Device{}, fmt.Errorf("patch device: %w", err)
-	}
-	return d, nil
-}
-
